+++ admin/src/components/edit-tag/edit-tag.jsx
@@ -3,7 +3,7 @@ import { useParams, Redirect } from "react-router";
 import { Container, Form, Button } from "react-bootstrap";
 import { useHistory } from "react-router-dom";
 import { useIntl, FormattedMessage } from "react-intl";
-import getFormErrors from "../util/form-errors-helper";
+import getFormErrors from "../util/helpers/form-errors-helper";
 import FormInput from "../util/forms/form-input";
 
 /**
